# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Test Academy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mcq/levels/{levelId}/questions:
    get:
      operationId: listMcqQuestions
      tags: [mcq]
      summary: List questions for a level
      parameters:
        - in: path
          name: levelId
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/McqQuestion"
    post:
      operationId: createMcqQuestion
      tags: [mcq]
      summary: Create MCQ question (admin)
      parameters:
        - in: path
          name: levelId
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McqQuestionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McqQuestion"
  /mcq/questions/{id}:
    patch:
      operationId: updateMcqQuestion
      tags: [mcq]
      summary: Update MCQ question (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McqQuestionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McqQuestion"
    delete:
      operationId: deleteMcqQuestion
      tags: [mcq]
      summary: Delete MCQ question (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
  /mcq/submit:
    post:
      operationId: submitMcqAnswers
      tags: [mcq]
      summary: Submit MCQ answers for a level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/McqSubmission"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/McqResult"
...
  /physical:
    get:
      operationId: listPhysical
      tags: [physical]
      summary: List physical guidelines
      parameters:
        - in: query
          name: all
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
      responses:
        "200":
          description: List of physical guidelines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PhysicalItem"
    post:
      operationId: createPhysical
      tags: [physical]
      summary: Create physical guideline (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhysicalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhysicalItem"
  /physical/{id}:
    patch:
      operationId: updatePhysical
      tags: [physical]
      summary: Update physical guideline (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhysicalUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhysicalItem"
    delete:
      operationId: deletePhysical
      tags: [physical]
      summary: Delete physical guideline (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
  /physical/{id}/publish:
    patch:
      operationId: publishPhysical
      tags: [physical]
      summary: Toggle publish physical guideline (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhysicalItem"
...
components:
  schemas:
...
    McqQuestion:
      type: object
      required: [id, levelId, question, optionA, optionB, optionC, optionD, correctAnswer, orderIndex]
      properties:
        id:
          type: number
        levelId:
          type: number
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          type: string
        orderIndex:
          type: number
    McqQuestionInput:
      type: object
      required: [question, optionA, optionB, optionC, optionD, correctAnswer]
      properties:
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          $ref: "#/components/schemas/McqQuestionInputCorrectAnswer"
        orderIndex:
          type: number
    McqQuestionInputCorrectAnswer:
      type: string
      enum: [A, B, C, D]
    McqQuestionUpdate:
      type: object
      properties:
        question:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctAnswer:
          $ref: "#/components/schemas/McqQuestionUpdateCorrectAnswer"
        orderIndex:
          type: number
    McqQuestionUpdateCorrectAnswer:
      type: string
      enum: [A, B, C, D]
...
    McqAnswer:
      type: object
      required: [questionId, selectedAnswer]
      properties:
        questionId:
          type: number
        selectedAnswer:
          $ref: "#/components/schemas/McqAnswerSelectedAnswer"
    McqAnswerSelectedAnswer:
      type: string
      enum: [A, B, C, D]
    McqSubmission:
      type: object
      required: [levelId, answers]
      properties:
        levelId:
          type: number
        answers:
          type: array
          items:
            $ref: "#/components/schemas/McqAnswer"
    McqResult:
      type: object
      required: [score, total, percentage]
      properties:
        score:
          type: number
        total:
          type: number
        percentage:
          type: number
...
    PhysicalItem:
      type: object
      required: [id, category, title, content, isPublished, createdAt, updatedAt]
      properties:
        id:
          type: number
        category:
          type: string
        title:
          type: string
        content:
          type: string
        isPublished:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    PhysicalInput:
      type: object
      required: [category, title, content]
      properties:
        category:
          type: string
        title:
          type: string
        content:
          type: string
        isPublished:
          type: boolean
    PhysicalUpdate:
      type: object
      properties:
        category:
          type: string
        title:
          type: string
        content:
          type: string
...
    PublishInput:
      type: object
      required: [isPublished]
      properties:
        isPublished:
          type: boolean
...