# lib/api-zod/src/generated/types/mcqAnswerSelectedAnswer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Test Academy API
 * OpenAPI spec version: 0.1.0
 */

export type McqAnswerSelectedAnswer = typeof McqAnswerSelectedAnswer[keyof typeof McqAnswerSelectedAnswer];


export const McqAnswerSelectedAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;