# lib/api-zod/src/generated/types/mcqQuestionInputCorrectAnswer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Test Academy API
 * OpenAPI spec version: 0.1.0
 */

export type McqQuestionInputCorrectAnswer = typeof McqQuestionInputCorrectAnswer[keyof typeof McqQuestionInputCorrectAnswer];


export const McqQuestionInputCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;