# lib/api-zod/src/generated/types/mcqQuestionUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Test Academy API
 * OpenAPI spec version: 0.1.0
 */
import type { McqQuestionUpdateCorrectAnswer } from './mcqQuestionUpdateCorrectAnswer';

export interface McqQuestionUpdate {
  question?: string;
  optionA?: string;
  optionB?: string;
  optionC?: string;
  optionD?: string;
  correctAnswer?: McqQuestionUpdateCorrectAnswer;
  orderIndex?: number;
}