# lib/api-zod/src/generated/types/mcqQuestionUpdateCorrectAnswer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Test Academy API
 * OpenAPI spec version: 0.1.0
 */

export type McqQuestionUpdateCorrectAnswer = typeof McqQuestionUpdateCorrectAnswer[keyof typeof McqQuestionUpdateCorrectAnswer];


export const McqQuestionUpdateCorrectAnswer = {
  A: 'A',
  B: 'B',
  C: 'C',
  D: 'D',
} as const;